wifi: Allow managing system-level connections
This is necessary to connect to a new network, or to forget a known
network.

Using system-level connections rather than user-level connections means
that NetworkManager will automatically connect to the network at
boot-time, without needing a user session to be open first.

This is necessary because the ideascube application runs as a service,
so that even if it runs as the ideascube user, it is not part of a user
session.