Use a PAM session for uwsgi
The uwsgi service runs the ideascube application as the ideascube user.

The ideascube application tries to talk to NetworkManager to manage the
Wi-Fi networks on the box.

We already gave the ideascube user the proper permissions to the right
NetworkManager actions through PolicyKit.

Unfortunately, on top of the PolicyKit permissions, NetworkManager also
requires the client process to be part of a user session.

This commit achieves that.