Complete upgrade
We use pip to install ka-lite, this is the best way found so far. The
installed version is given from local fact as we need to control the
version and don't want always the last version...

During operation, kalite has to be stopped and started, we use therefore
wait_for task to be sure the service is started or stopped.

Everything regarding the storage of the media on an external hard drive
have been moved to the file ext_hdd.yml