Upgrade from old AnsibleCube version
This file and playbook are used to upgrade from the old AnsibleCube to
the new AnsibleCube.

Each update_*.yml file will be simlink to upgrade.yml. The role
upgrade_device will gather all the software version installed on the
device and stored as fact. Once this will be done, a new ansible-pull
file will be copied in
/etc/NetworkManager/dispatcher.d/ansiblePullUpdate to call the main.yml
file. The rest will fellow