Clean up old ZIM install for IdeasBox server
This role aim to move all ZIM present in /usr/local/share/kiwix to
the new Kiwix folder in /var/ideascube/kiwix.

To do so we look for all the zim in old path, create a library.xml file
from each zim, compress the data in an archive, get a new checksum
from the latter and then create a catalog.yml from all the archive.
At this stage all the archive are moved to /var/cache/ideascube/packages
and are ready to processed by ideascube catalog.

At the end we have a clean collection of zim in our new zim folder.
Index folder are not yet copied, but most of the time they were absent
from the configuration.