Main playbook : Move roles around
Move up zim_install role and move down kalite install.

Indeed, zim_install is requiering a lot of space to download and
uncompress zim archive. Sometime we are missing space to proced to this
task so we do it first

Kalite role is just about copying file from a NAS to the device and don't
requier more space drive to handle this task