Change way of pushing the logs
AnsibleCube was launch each time an Internet connection was detected mainly
because the latter was collecting and pushing logs to the central
server.

The whole update run consume a lot of ressources on the device and it is
preferable to execute the play-book only when the github repository has
changed.

This commit split the pushing log action and copy the push_log bash
script in the NetworkManager dispastech.d folder to be executed each
time the device get connected to Internet