Set the right IDEASCUBE_ID
We are moving towards Ideascube configuration file free. But still we
need sometime a configuration file for cards that are not automatically
enabled.

This commit looks for a configuration file in the Ideascube configuration
file folder. If it exist the latter is used, otherwise we are falling
back to the generic configuration for all KoomBook : kb.py

If a device needs a non automatic card enabled on the homepage, we will have to create a
specific Ideascube configuration file to specify what card has to be
shown.

So far we still need that the name of the Ideascube configuration file match the
name of the AnsibleCube device_lists.fact file otherwise kb.py will be
used