Move fact registration to set_custom_fact playbook
Some facts has to be register at the early beginning of the playbook.
It is also more convenient to have them all in the same place.

Tags have been introduce in this playbook because all the tasks don't have to
run with the same tag

This change allow us to register the new variable own_config_file that
tell if the device_list.fact file has to be copied