Include ext_hdd taks after installing ideascube
This bug has been discovered while installing a KoomBook with master and
custom tags, then with an update tag.

Because we are using an external hard drive, we have to first install
ideascube then do the appropriate process to move the ideascube folder
to the hard drive. Because we were mixing master and custom tag, the
ideascube folder was not moved to the hard drive, this task was done at
the update tag, but with the wrong sequence.

At this stage we are able to properly run the update
stage without any error at the ideascube level.