change hosts file if koombook-doctor answer ping
This script is executed by network manager with the dispatcher.

If koombook-doctor.local is present in the local network, look for it
IP address then add it to /etc/hosts file. Then all requests to
download.kiwix.org and other website will be answered by the local cache
(koombook-doctor)

This way of adding the IP address to /etc/hosts (with a bash script) file is better than with
ansible, as we may have admin using directly the `ideascube catalog
install wikiepedia.en` instead of letting ansible dealing with the
downloading. This way we ensure that our server is all the time
checking for a koombook-doctor.local