Install, upgrade or remove a package
Fellowing the value given in each status package, install, upgrade or
remove the latter.

Jinja filter has been used to select the right command to be done.
Since the command "ideascube catalog install package" just install the
package, for a package update, in all case,  we have first to install the package then
update it. It would be easier if ideascube catalog install could install
and upgrade the package if the latter is already installed.