Simplify handling unsigned packages
Only our Ideascube package should be unsigned, and it is the only one we
want to trust.

This sets things up so that only the packages coming from our repository
are trusted, whereas we used to disable the verification of the
signatures for all the packages installed at the same time as Ideascube,
regardless of where they came from.