Enable the Ideascube web app towards the end
I have no clue why we were doing this uWSGI symlink in the first place,
since the ideascube package was already doing it.

However, the ideascube package stopped doing it with 0.29.2, both for
uWSGI and Nginx, so we definitely need to do it for both of them now.

We shouldn't enable the apps before everything is ready, though.