Run the Ideascube migrations on initial installation
We do not start Ideascube right after installing it any more, now that
the application isn't enabled in uWSGI by the postinst script in the
package.

As a result, the migrations, which are run automatically at startup,
have never been run yet.

Since they are necessary for the various `ideascube config set ...`
calls, let's run them manually.

However, we should only run them on initial installation, never when
updating the package. On updates, we want to rely on the automatic run
done when restarting the application.