Add a device "KoomBook"
I create a new device named KoomBook, this does not match our
nomenclature but it will be used to create a new generic KoomBook image
that could be renamed latter on. This is specially usefull for KoomBook
DIY workshop where our main focus is not on software and content but
hardware. The script
https://github.com/ideascube/ansiblecube/blob/oneUpdateFile/tools/koombook_zeroconf
will be automatically executed at first boot.