Help end-user by doing no conf on KB before using it
This script has to be drop in /etc/NetworkManager/dispatcher.d/ with
chmod 755
It works along the KoomBook image generated by compilBook
The purpose of this script is to automate the configuration of a brand
new
assembled KoomBook.
It will detect the main external HDD /dev/sda and create a new partition
on ext4 format.
Mount the newly partition and move the ideascube data folder to the
external HDD.
The last operation will give a new name to the device