Factorize the `restart uwsgi` handler
Having it in each and every role means that Ansible considers them
separate handlers.

As a result, uWSGI can be restarted more than once at the end of a
playbook.

When it starts, uWSGI runs the Ideascube app, which runs collectstatic.

Because the uWSGI service is `Type=forking`, systemd considers uWSGI
finished started when the parent process exits. But that happens long
before collectstatic finished running.

As a result, it is possible that Ansible will restart uWSGI twice in a
row, while the first one was still running collectstatic.

We think this is what could cause a weird Ideascube bug:

https://framagit.org/ideascube/ideascube/issues/829