Disabling faulty include.
Including dust-removal.yml which has a task calling for a fake catalog kiwix.yml, but located on a remote server. The file is gone, no backup, we just don't know what was its purpose. Waiting for @fheslouin to give us a hint. In the meantime, just don't do dust-removal and let the thing dirty a bit longer.