Enable use of external antenna
This new tasks use a script to detect if the device got an IP address
given by a dhcp server,
if true, the device is within a local network probably connected to
Internet. If the device does not get any IP addresss but an ethernet
link is detected we asign an arbitrary IP address to the interface and
we assume that an external Ubiquiti wifi antenna is connected and is
waiting for an IP address from given by our device.