Split logs_to_hdd tasks
I had to move some ansible tasks to a bash script executed at the first
boot. The main reason is that the HDD might not be connected when the
master tag is executed. Thus, HDD is mounted when the device start for
the first start, script koombook_zeroconf take care of this, it then
synchronise the log folder and so on.